rename spec classes to fit with #143
